requirements.txt
# Pantry - Python Dependencies

# Web Framework
streamlit>=1.30.0

# Database
psycopg2-binary>=2.9.0

# Data Processing
pandas>=2.1.0
beautifulsoup4>=4.12.0

# AI/LLM
openai>=1.0.0

# Browser Automation
playwright>=1.40.0

# Image Processing
Pillow>=10.0.0

# Environment
python-dotenv>=1.0.0

# HTTP
requests>=2.31.0
